+++ src/components/Microsoft.DotNet.UpgradeAssistant.MSBuild/MSBuildProject.cs
@@ -9,8 +9,6 @@ using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.CodeAnalysis;
 using Microsoft.Extensions.Logging;
-using NuGet.Frameworks;
-using NuGet.ProjectModel;
 
 using MBuild = Microsoft.Build.Evaluation;
 
@@ -102,93 +100,6 @@ namespace Microsoft.DotNet.UpgradeAssistant.MSBuild
             }
         }
 
-        public NugetPackageFormat PackageReferenceFormat
-        {
-            get
-            {
-                if (GetPackagesConfigPath() is not null)
-                {
-                    return NugetPackageFormat.PackageConfig;
-                }
-                else if (ProjectRoot.GetAllPackageReferences().ToList() is IEnumerable<Build.Construction.ProjectItemElement> list && list.Any())
-                {
-                    return NugetPackageFormat.PackageReference;
-                }
-                else
-                {
-                    return NugetPackageFormat.None;
-                }
-            }
-        }
-
-        private string? GetPackagesConfigPath() => FindFiles(ProjectItemType.Content, "packages.config").FirstOrDefault();
-
-        public IEnumerable<NuGetReference> PackageReferences
-        {
-            get
-            {
-                var packagesConfig = GetPackagesConfigPath();
-
-                if (packagesConfig is null)
-                {
-                    var packages = ProjectRoot.GetAllPackageReferences();
-
-                    return packages.Select(p => p.AsNuGetReference()).ToList();
-                }
-                else
-                {
-                    return PackageConfig.GetPackages(packagesConfig);
-                }
-            }
-        }
-
-        public IEnumerable<NuGetReference> GetTransitivePackageReferences(TargetFrameworkMoniker tfm)
-        {
-            if (PackageReferenceFormat != NugetPackageFormat.PackageReference)
-            {
-                return Enumerable.Empty<NuGetReference>();
-            }
-
-            var parsedTfm = NuGetFramework.Parse(tfm.Name);
-            var lockFile = LockFileUtilities.GetLockFile(LockFilePath, NuGet.Common.NullLogger.Instance);
-
-            if (lockFile is null)
-            {
-                throw new ProjectRestoreRequiredException($"Project is not restored: {FileInfo}");
-            }
-
-            var lockFileTarget = lockFile
-                .Targets
-                .FirstOrDefault(t => t.TargetFramework.DotNetFrameworkName.Equals(parsedTfm.DotNetFrameworkName, StringComparison.Ordinal));
-
-            if (lockFileTarget is null)
-            {
-                throw new ProjectRestoreRequiredException($"Could not find {parsedTfm.DotNetFrameworkName} in {LockFilePath} for {FileInfo}");
-            }
-
-            return lockFileTarget.Libraries.Select(l => new NuGetReference(l.Name, l.Version.ToNormalizedString()));
-        }
-
-        public string? LockFilePath
-        {
-            get
-            {
-                var lockFilePath = Path.Combine(GetPropertyValue("MSBuildProjectExtensionsPath"), "project.assets.json");
-
-                if (string.IsNullOrEmpty(lockFilePath))
-                {
-                    return null;
-                }
-
-                if (!Path.IsPathFullyQualified(lockFilePath))
-                {
-                    lockFilePath = Path.Combine(FileInfo.DirectoryName ?? string.Empty, lockFilePath);
-                }
-
-                return lockFilePath;
-            }
-        }
-
         public IEnumerable<Reference> FrameworkReferences =>
             ProjectRoot.GetAllFrameworkReferences().Select(r => r.AsReference()).ToList();
 
@@ -196,8 +107,6 @@ namespace Microsoft.DotNet.UpgradeAssistant.MSBuild
             ProjectRoot.GetAllReferences().Select(r => r.AsReference()).ToList();
 
         public IReadOnlyCollection<TargetFrameworkMoniker> TargetFrameworks => new TargetFrameworkMonikerCollection(this);
-
-        IProjectFile IProject.GetFile() => this;
 
         public override bool Equals(object? obj)
         {
